Fixed tests and added a couple of suggestions
It was reported in Slack that shot_rspec tests weren't passing despite
that they were pretty simple.

I've made a couple of changes in order to make the tests to pass and
also made some code-style changes in order to make it a bit more clean.

Suggestions:
- Check your indentation, it makes easier to read code when you follow
  conventions. :D
- Ruby always return the last operation that was executed, so, even when
  the `get_type` method sets the right value to `type`, it's needed to
  return it explictly at the end of the method
- Consider those scenarios in which the sum of both shoots returns 10,
  even when the first one is 10 and the second one 0, since it's running
  sequentially, there is nothing that avoids classifying it as a spare.

  Consider making an early return when any of the consitions is matched.
- It was missed the arg for `is_normal?``

Also, I'd recommend using namend-args rather than passing a single
argument with a hash. While it's simple, you can avoid passing the wrong
hash-struct. Just a thing to keep in mind